Rename client module
