+++ lib/client.js
