+++ lib/interactions.js
@@ -1,5 +1,5 @@
 import { APIMessage, MessageEmbed } from 'discord.js'
-import { getClient } from './discord-client.js'
+import { getClient } from './client.js'
 import { SUB_COMMAND, SUB_COMMAND_GROUP, INTEGER, CHANNEL_MESSAGE_WITH_SOURCE } from './constants.js'
 import { debug, info } from './logger.js'
 import { roll, getResult } from './dice.js'
